Various tweaks to increase perf of bloom filter / rabin-karp hashing
In particular it uses 64bit uint hashing now, also uses 5 hash functions
instead of 3 to reduce the number of false positives (which directly
affects perf).  Also figures out better fingerprints for the .dat file.

+++ Application/LinkBubble/src/main/jni/hashFn.h
@@ -9,13 +9,13 @@ public:
     this->p = p;
   }
 
-  virtual int operator()(const char *input, int len, unsigned char lastCharCode, unsigned int lastHash) {
+  virtual uint64_t operator()(const char *input, int len, unsigned char lastCharCode, uint64_t lastHash) {
     // See the abracadabra example: https://en.wikipedia.org/wiki/Rabin%E2%80%93Karp_algorithm
     return (lastHash - lastCharCode * pow(p, len - 1)) * p + input[len - 1];
   }
 
-  virtual int operator()(const char *input, int len) {
-    int total = 0;
+  virtual uint64_t operator()(const char *input, int len) {
+    uint64_t total = 0;
     for (int i = 0; i < len; i++) {
       total += input[i] * pow(p, len - i - 1);
     }
